+++ src/main/java/kittify/common/CommonEventHandler.java
@@ -8,6 +8,7 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.entity.item.EntityXPOrb;
 import net.minecraft.entity.monster.EntityCreeper;
 import net.minecraft.entity.passive.EntityTameable;
 import net.minecraft.entity.player.EntityPlayer;
@@ -15,9 +16,12 @@ import net.minecraft.util.DamageSource;
 import net.minecraft.world.Explosion;
 import net.minecraft.world.World;
 import net.minecraftforge.event.entity.living.LivingAttackEvent;
+import net.minecraftforge.event.entity.living.LivingExperienceDropEvent;
 import net.minecraftforge.event.entity.living.LivingSetAttackTargetEvent;
+import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.event.world.ExplosionEvent;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
 
@@ -123,6 +127,39 @@ public class CommonEventHandler {
             explosion.clearAffectedBlockPositions();
             // No damage to non-living entities
             e.getAffectedEntities().removeIf(entity -> !(entity instanceof EntityLiving));
+        }
+    }
+
+    /**
+     * Drop all experience on death.
+     * Handles with lowest priority so that other mods can do stuff with this event if they want to first.
+     * This might cause a whole lot of lag though... Haha... Might wanna install clumps mod tbh.
+     * This doesn't get run if keepInventory is on.
+     * @param e Event data
+     */
+    @SubscribeEvent(priority = EventPriority.LOWEST)
+    public static void dropAllExperienceOnDeath(LivingExperienceDropEvent e) {
+        final EntityLivingBase entity = e.getEntityLiving();
+        if(entity instanceof EntityPlayer) {
+            final EntityPlayer player = (EntityPlayer) entity;
+            final int experienceTotal = player.experienceTotal;
+            if(e.getDroppedExperience() < experienceTotal) {
+                // Handle spawning experience specially to cause less lag
+                e.setDroppedExperience(0);
+                e.setCanceled(true);
+                int remaining = experienceTotal;
+                // Experience orbs have a limit of short value (halving just in case?)
+                final int maxExpValue = Short.MAX_VALUE / 2;
+                final World world = player.world;
+                final double posX = player.posX;
+                final double posY = player.posY;
+                final double posZ = player.posZ;
+                while(remaining > 0) {
+                    final int splitExp = Math.min(maxExpValue, remaining);
+                    remaining -= splitExp;
+                    world.spawnEntity(new EntityXPOrb(world, posX, posY, posZ, splitExp));
+                }
+            }
         }
     }
 }